+++ glean-core/src/error_recording.rs
@@ -13,13 +13,17 @@
 //! not some constant value that we could define in `metrics.yaml`.
 
 use std::fmt::Display;
+#[cfg(feature = "sqlite")]
 use std::sync::atomic::AtomicU8;
 
+#[cfg(feature = "sqlite")]
 use rusqlite::Transaction;
 
 use crate::common_metric_data::CommonMetricDataInternal;
 use crate::error::{Error, ErrorKind};
-use crate::metrics::{CounterMetric, Metric};
+use crate::metrics::CounterMetric;
+#[cfg(feature = "sqlite")]
+use crate::metrics::Metric;
 use crate::Glean;
 use crate::Lifetime;
 use crate::{CommonMetricData, MetricLabel};
@@ -139,12 +143,13 @@ pub fn record_error<O: Into<Option<i32>>>(
 ) {
     let metric = get_error_metric_for_metric(meta, error);
 
-    log::warn!("{}: {}", meta.base_identifier(), message);
+    log::warn!("{}: {})", meta.base_identifier(), message);
     let to_report = num_errors.into().unwrap_or(1);
     debug_assert!(to_report > 0);
     metric.add_sync(glean, to_report);
 }
 
+#[cfg(feature = "sqlite")]
 pub fn record_error_sqlite(
     glean: &Glean,
     tx: &mut Transaction,