+++ glean-core/src/internal_metrics.rs
@@ -361,6 +361,9 @@ pub struct DatabaseMetrics {
     /// sqlite's load result, indicating success or relaying the detected error.
     pub load_error: StringMetric,
 
+    /// Rkv's load result, indicating success or relaying the detected error.
+    pub rkv_load_error: StringMetric,
+
     /// The time it takes for a write-commit for the Glean database.
     pub write_time: TimingDistributionMetric,
 
@@ -397,6 +400,14 @@ impl DatabaseMetrics {
 
             load_error: StringMetric::new(CommonMetricData {
                 name: "load_error".into(),
+                category: "glean.database".into(),
+                send_in_pings: vec!["metrics".into(), "health".into()],
+                lifetime: Lifetime::Ping,
+                ..Default::default()
+            }),
+
+            rkv_load_error: StringMetric::new(CommonMetricData {
+                name: "rkv_load_error".into(),
                 category: "glean.database".into(),
                 send_in_pings: vec!["metrics".into(), "health".into()],
                 lifetime: Lifetime::Ping,