+++ glean-core/src/metrics/string_list.rs
@@ -154,7 +154,12 @@ impl StringListMetric {
             .into()
             .unwrap_or_else(|| &self.meta().inner.send_in_pings[0]);
 
-        match glean.storage().get_metric(self.meta(), queried_ping_name) {
+        match glean.storage().get_metric(
+            #[cfg(not(feature = "sqlite"))]
+            glean,
+            self.meta(),
+            queried_ping_name,
+        ) {
             Some(Metric::StringList(values)) => Some(values),
             _ => None,
         }