+++ glean-core/src/metrics/text.rs
@@ -99,7 +99,12 @@ impl TextMetric {
             .into()
             .unwrap_or_else(|| &self.meta().inner.send_in_pings[0]);
 
-        match glean.storage().get_metric(self.meta(), queried_ping_name) {
+        match glean.storage().get_metric(
+            #[cfg(not(feature = "sqlite"))]
+            glean,
+            self.meta(),
+            queried_ping_name,
+        ) {
             Some(Metric::Text(s)) => Some(s),
             _ => None,
         }